+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Parameter class hierarchy
-- Searchspace has now a discrete and continuous subspace
+- SearchSpace has now a discrete and continuous subspace
 
 ## [0.2.2] - 2023-01-13
 ### Added